refactor: move temporal data schema type enum
