+++ common/src/main/kotlin/org/neo4j/connectors/kafka/data/TemporalDataSchemaType.kt
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.neo4j.connectors.kafka
+package org.neo4j.connectors.kafka.data
 
 enum class TemporalDataSchemaType(val description: String) {
   STRUCT("struct"),
