+++ common/src/test/kotlin/org/neo4j/connectors/kafka/data/DynamicTypesTest.kt
@@ -40,7 +40,6 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.ArgumentsProvider
 import org.junit.jupiter.params.provider.ArgumentsSource
-import org.neo4j.connectors.kafka.TemporalDataSchemaType
 import org.neo4j.driver.Value
 import org.neo4j.driver.Values
 import org.neo4j.driver.types.Node
